refactor: 디스패처 서블릿 초기 설정 책임 분리
