+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/DispatcherServlet.java
@@ -6,43 +6,27 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.adapter.Adapter;
 import webmvc.org.springframework.web.servlet.mvc.adapter.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.mvc.adapter.HandlerExecutionAdapter;
 import webmvc.org.springframework.web.servlet.mvc.mapper.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.mapper.Mapper;
-import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class DispatcherServlet extends HttpServlet {
 
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private HandlerMapping handlerMapping;
-    private HandlerAdapter handlerAdapter;
+    private final HandlerMapping handlerMapping;
+    private final HandlerAdapter handlerAdapter;
+
+    public DispatcherServlet(final HandlerMapping handlerMapping, final HandlerAdapter handlerAdapter) {
+        this.handlerMapping = handlerMapping;
+        this.handlerAdapter = handlerAdapter;
+    }
 
     @Override
     public void init() {
-        handlerMapping = handlerMapping();
         handlerMapping.initialize();
-        handlerAdapter = handlerAdapter();
     }
 
-
-    private HandlerMapping handlerMapping() {
-        List<Mapper> mappers = new ArrayList<>();
-        mappers.add(new AnnotationHandlerMapping("com.techcourse.controller"));
-        return new HandlerMapping(mappers);
-    }
-
-    private HandlerAdapter handlerAdapter() {
-        List<Adapter> adapters = new ArrayList<>();
-        adapters.add(new HandlerExecutionAdapter());
-        return new HandlerAdapter(adapters);
-    }
 
     @Override
     protected void service(final HttpServletRequest request, final HttpServletResponse response) throws ServletException {
